Add hours duration argument
We currently support minutes and seconds, so let's also support hours.
Unfortunately the natural flag to use '-h' is already taken by 'help',
so I'm deciding to use '-H' instead. An argument could be made for
consistency that the other flags should be made capital to, but I'll
have to think more about it.