Clean up logging a little
Debugging is probably not all that helpful to users not developing this
tool, so let's set our level to INFO by default.

Also simplify the format a bit by removing function (we're not really
using any in the current implementation anyway) and putting the level
closer to the message, which reads a bit clearer.